searchlens-backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
numpy
